+++ src/components/validation/resetvalidation.ts
@@ -1,0 +1,10 @@
+import * as Yup from "yup";
+
+export const validationSchema = Yup.object({
+  password: Yup.string()
+    .min(8, "Password must be at least 8 characters long")
+    .required("Required"),
+  confirmPassword: Yup.string()
+    .oneOf([Yup.ref('password')], "Passwords must match")
+    .required("Required"),
+});
